Brian's Edits
Brian Moran from Ventric edited this file so that when you use it as a
post processor, and you have a spindle speed set for a tool that is
from 1-99 the spindle (or router) will never be activated. This is to
allow for the ShopBot to be used to slide parts around in the workspace
without having to manually edit the scripts.